clustering

+++ nemo/collections/asr/data/feature_to_label.py
@@ -78,7 +78,9 @@ class _FeatureSeqSpeakerLabelDataset(Dataset):
             )
         else:
             output_types.update(
-                {'label': NeuralType(('B', 'T'), LabelsType()), 'label_length': NeuralType(tuple('B'), LengthsType()),}
+                {
+                    'label': NeuralType(('B', 'T'), LabelsType()), 
+                    'label_length': NeuralType(tuple('B'), LengthsType()),}
             )
 
         return output_types
@@ -110,7 +112,7 @@ class _FeatureSeqSpeakerLabelDataset(Dataset):
         features = self.feature_loader.process(sample.feature_file)
         f, fl = features, torch.tensor(features.shape[0]).long()
 
-        t = torch.tensor(sample.seq_label).float()
+        t = torch.tensor(sample.seq_label).long()
         tl = torch.tensor(len(sample.seq_label)).long()
 
         return f, fl, t, tl
